Checking error only for pipes without quotes
