+++ Makefile
@@ -50,7 +50,8 @@ SRCS	:=	./src/main.c \
 			./src/builtins/ft_cd.c \
 			./src/builtins/ft_exit.c \
 			./src/print_array.c \
-			./src/split_processes.c
+			./src/split_processes.c \
+			./src/error_syntax_pipes.c
 			
 OBJS	:= ${SRCS:.c=.o}
